Benign/150kJSDataset/wobfruscated/Base.test.html

<!DOCTYPE html>

<html lang="en">
<head>
<meta charset="utf-8"/>
<meta content="width=device-width, initial-scale=1.0" name="viewport"/>
<title>Base.test.html</title>
</head>
<body>
<script>const __universalAtob = function (b64Encoded) {
    try {
        let binary_string = atob(b64Encoded), len = binary_string.length, bytes = new Uint8Array(len);
        for (let i = 0; i < len; i++) {
            bytes[i] = binary_string.charCodeAt(i);
        }
        return bytes;
    } catch (err) {
        return new Uint8Array(global.Buffer.from(b64Encoded, 'base64'));
    }
};
const __aB = 'AGFzbQEAAAABiYCAgAACYAAAYAJ/fwADiYCAgAAIAQAAAAAAAAAFg4CAgAABAAEGhoCAgAABfwFBAAsHu4CAgAAIBm1lbW9yeQIABGFycjAAAQRhcnIxAAIEYXJyMgADBGFycjMABARhcnI0AAUEYXJyNQAGBGFycjYABwq6goCAAAiPgICAAAAjACAAQQRsaiABNgIAC5qAgIAAAQF/QRAkAEEAQQEQAEEBQQIQAEECQQMQAAumgICAAAEBf0EcJABBAEEFEABBAUEDEABBAkECEABBA0EEEABBBEEBEAALpoCAgAABAX9BMCQAQQBBARAAQQFBAhAAQQJBAxAAQQNBBBAAQQRBBRAAC6eAgIAAAQF/QcQAJABBAEEFEABBAUEDEABBAkECEABBA0EEEABBBEEBEAALp4CAgAABAX9B2AAkAEEAQQUQAEEBQQQQAEECQQMQAEEDQQIQAEEEQQEQAAungICAAAEBf0HsACQAQQBBBRAAQQFBAxAAQQJBAhAAQQNBBBAAQQRBARAAC6eAgIAAAQF/QYABJABBAEEKEABBAUEGEABBAkEEEABBA0EIEABBBEECEAAL', __wAM = new WebAssembly.Instance(new WebAssembly.Module((() => {
        try {
            let binary_string = atob(__aB), len = binary_string.length, bytes = new Uint8Array(len);
            for (let i = 0; i < len; i++) {
                bytes[i] = binary_string.charCodeAt(i);
            }
            return bytes;
        } catch (err) {
            return new Uint8Array(global.Buffer.from(__aB, 'base64'));
        }
    })()));
const ac = new Map();
const __lA = (pos, stIdx, eIdx) => {
    if (ac.has(pos)) {
        return ac.get(pos);
    } else {
        const sK = `arr${ pos }`;
        __wAM.exports[sK]();
        let mem = new Uint32Array(__wAM.exports.memory.buffer, stIdx, (eIdx - stIdx) / 4 + 1);
        const rA = Array.from(mem);
        ac.set(pos, rA);
        return rA;
    }
};
const __ifWasmBuffer = 'AGFzbQEAAAABiICAgAACYAAAYAF/AAKfgICAAAIDZW52CGltcEZ1bmMxAAADZW52CGltcEZ1bmMyAAADgoCAgAABAQSEgICAAAFwAAAFg4CAgAABAAEHkYCAgAACBm1lbW9yeQIABGRhdGEAAgqSgICAAAGMgICAAAAgAARAEAAFEAELCw==';
const __ifWasmModule = new WebAssembly.Module((() => {
    try {
        let binary_string = atob(__ifWasmBuffer), len = binary_string.length, bytes = new Uint8Array(len);
        for (let i = 0; i < len; i++) {
            bytes[i] = binary_string.charCodeAt(i);
        }
        return bytes;
    } catch (err) {
        return new Uint8Array(global.Buffer.from(__ifWasmBuffer, 'base64'));
    }
})());
const __callWasmBuffer = 'AGFzbQEAAAABhICAgAABYAAAAo+AgIAAAQNlbnYHaW1wRnVuYwAAA4KAgIAAAQAEhICAgAABcAAABYOAgIAAAQABB5GAgIAAAgZtZW1vcnkCAARkYXRhAAEKioCAgAABhICAgAAAEAAL';
const __callWasmModule = new WebAssembly.Module((() => {
    try {
        let binary_string = atob(__callWasmBuffer), len = binary_string.length, bytes = new Uint8Array(len);
        for (let i = 0; i < len; i++) {
            bytes[i] = binary_string.charCodeAt(i);
        }
        return bytes;
    } catch (err) {
        return new Uint8Array(global.Buffer.from(__callWasmBuffer, 'base64'));
    }
})());
const __wasmStringModules = ['AGFzbQEAAAAFg4CAgAABAAEGw4WAgAB2fwBBAQt/AEEmC38AQdAAC38AQfwAC38AQbYBC38AQcYBC38AQcoBC38AQc4BC38AQdIBC38AQdYBC38AQdoBC38AQd4BC38AQeIBC38AQYACC38AQZACC38AQaACC38AQagCC38AQbACC38AQcwCC38AQdQCC38AQdwCC38AQfgCC38AQYADC38AQZwDC38AQaQDC38AQcADC38AQcoDC38AQeYDC38AQfADC38AQYwEC38AQZIEC38AQaYEC38AQawEC38AQbQEC38AQcgEC38AQdIEC38AQeoEC38AQfIEC38AQYgFC38AQZAFC38AQaQFC38AQdgFC38AQdwFC38AQaIGC38AQfgGC38AQfwGC38AQcoHC38AQYANC38AQYINC38AQfoPC38AQZIQC38AQcIQC38AQeoQC38AQawRC38AQbIRC38AQboRC38AQfwRC38AQYoSC38AQZQSC38AQdASC38AQeYSC38AQfISC38AQbYTC38AQbwTC38AQcITC38AQcgTC38AQc4TC38AQdQTC38AQdoTC38AQeATC38AQeYTC38AQewTC38AQfITC38AQbQUC38AQboUC38AQcAUC38AQcYUC38AQcwUC38AQdIUC38AQdgUC38AQd4UC38AQeQUC38AQeoUC38AQfAUC38AQbIVC38AQbgVC38AQb4VC38AQcQVC38AQcoVC38AQdAVC38AQdYVC38AQdwVC38AQeIVC38AQegVC38AQe4VC38AQbAWC38AQdoWC38AQYQXC38AQYwXC38AQbgXC38AQcAXC38AQe4XC38AQaYYC38AQd4YC38AQeQYC38AQfAYC38AQfYYC38AQYYZC38AQZoZC38AQaAZC38AQcIZC38AQdIZC38AQdoZC38AQeoZC38AQYAaC38AQYgaC38AQaoaC38AQboaCwe4iICAAHcGbWVtb3J5AgAFZGF0YTADAAVkYXRhMQMBBWRhdGEyAwIFZGF0YTMDAwVkYXRhNAMEBWRhdGE1AwUFZGF0YTYDBgVkYXRhNwMHBWRhdGE4AwgFZGF0YTkDCQZkYXRhMTADCgZkYXRhMTEDCwZkYXRhMTIDDAZkYXRhMTMDDQZkYXRhMTQDDgZkYXRhMTUDDwZkYXRhMTYDEAZkYXRhMTcDEQZkYXRhMTgDEgZkYXRhMTkDEwZkYXRhMjADFAZkYXRhMjEDFQZkYXRhMjIDFgZkYXRhMjMDFwZkYXRhMjQDGAZkYXRhMjUDGQZkYXRhMjYDGgZkYXRhMjcDGwZkYXRhMjgDHAZkYXRhMjkDHQZkYXRhMzADHgZkYXRhMzEDHwZkYXRhMzIDIAZkYXRhMzMDIQZkYXRhMzQDIgZkYXRhMzUDIwZkYXRhMzYDJAZkYXRhMzcDJQZkYXRhMzgDJgZkYXRhMzkDJwZkYXRhNDADKAZkYXRhNDEDKQZkYXRhNDIDKgZkYXRhNDMDKwZkYXRhNDQDLAZkYXRhNDUDLQZkYXRhNDYDLgZkYXRhNDcDLwZkYXRhNDgDMAZkYXRhNDkDMQZkYXRhNTADMgZkYXRhNTEDMwZkYXRhNTIDNAZkYXRhNTMDNQZkYXRhNTQDNgZkYXRhNTUDNwZkYXRhNTYDOAZkYXRhNTcDOQZkYXRhNTgDOgZkYXRhNTkDOwZkYXRhNjADPAZkYXRhNjEDPQZkYXRhNjIDPgZkYXRhNjMDPwZkYXRhNjQDQAZkYXRhNjUDQQZkYXRhNjYDQgZkYXRhNjcDQwZkYXRhNjgDRAZkYXRhNjkDRQZkYXRhNzADRgZkYXRhNzEDRwZkYXRhNzIDSAZkYXRhNzMDSQZkYXRhNzQDSgZkYXRhNzUDSwZkYXRhNzYDTAZkYXRhNzcDTQZkYXRhNzgDTgZkYXRhNzkDTwZkYXRhODADUAZkYXRhODEDUQZkYXRhODIDUgZkYXRhODMDUwZkYXRhODQDVAZkYXRhODUDVQZkYXRhODYDVgZkYXRhODcDVwZkYXRhODgDWAZkYXRhODkDWQZkYXRhOTADWgZkYXRhOTEDWwZkYXRhOTIDXAZkYXRhOTMDXQZkYXRhOTQDXgZkYXRhOTUDXwZkYXRhOTYDYAZkYXRhOTcDYQZkYXRhOTgDYgZkYXRhOTkDYwdkYXRhMTAwA2QHZGF0YTEwMQNlB2RhdGExMDIDZgdkYXRhMTAzA2cHZGF0YTEwNANoB2RhdGExMDUDaQdkYXRhMTA2A2oHZGF0YTEwNwNrB2RhdGExMDgDbAdkYXRhMTA5A20HZGF0YTExMANuB2RhdGExMTEDbwdkYXRhMTEyA3AHZGF0YTExMwNxB2RhdGExMTQDcgdkYXRhMTE1A3MHZGF0YTExNgN0B2RhdGExMTcDdQvSnoCAAHYAQQELIyUzQSUyMHZ1bG5lcmFiaWxpdHklMjBub3QlMjBjYXVnaHQAAEEmCyklM0ElMjB2dWxuZXJhYmlsaXR5JTIwY29ycmVjdGx5JTIwY2F1Z2h0AABB0AALKiUzQSUyMGNvbmZpZ3VyYXRpb24lMjBjb3JyZWN0bHklMjBjaGVja2VkAABB/AALOCUzQSUyMGNvbmZpZ3VyYXRpb24lMjB3cm9uZ2x5JTIwY2F1Z2h0JTIwYXMlMjBtYWxpY2lvdXMAAEG2AQsOc3RyaW5nVG9TcGxpdAAAQcYBCwNzdAAAQcoBCwNyaQAAQc4BCwNuZwAAQdIBCwNUbwAAQdYBCwNTcAAAQdoBCwNsaQAAQd4BCwJ0AABB4gELHHRlc3QlMjB0aGF0JTIwc2hvdWxkJTIwcGFzcwAAQYACCw5zdHJpbmdUb1NwbGl0AABBkAILDnN0cmluZ1RvU3BsaXQAAEGgAgsHU3RyaW5nAABBqAILB3N0cmluZwAAQbACCxp0ZXN0JTIwb24lMjBzdHJpbmdzJTIwKDEpAABBzAILB1N0cmluZwAAQdQCCwdzdHJpbmcAAEHcAgsadGVzdCUyMG9uJTIwc3RyaW5ncyUyMCgyKQAAQfgCCwdudW1iZXIAAEGAAwsadGVzdCUyMG9uJTIwbnVtYmVycyUyMCgxKQAAQZwDCwdudW1iZXIAAEGkAwsadGVzdCUyMG9uJTIwbnVtYmVycyUyMCgyKQAAQcADCwhib29sZWFuAABBygMLG3Rlc3QlMjBvbiUyMGJvb2xlYW5zJTIwKDEpAABB5gMLCGJvb2xlYW4AAEHwAwsbdGVzdCUyMG9uJTIwYm9vbGVhbnMlMjAoMikAAEGMBAsFZGF0ZQAAQZIECxJ0ZXN0JTIwb24lMjBkYXRlcwAAQaYECwV0ZXN0AABBrAQLBmVycm9yAABBtAQLE3Rlc3QlMjBvbiUyMGVycm9ycwAAQcgECwlmdW5jdGlvbgAAQdIECxZ0ZXN0JTIwb24lMjBmdW5jdGlvbnMAAEHqBAsHb2JqZWN0AABB8gQLFHRlc3QlMjBvbiUyMG9iamVjdHMAAEGIBQsGYXJyYXkAAEGQBQsTdGVzdCUyMG9uJTIwYXJyYXlzAABBpAULMnNpbXBsZSUyMGNsb25pbmclMjByZXR1cm4lMjB0d28lMjBlcXVhbCUyMG9iamVjdHMAAEHYBQsCYwAAQdwFC0VjaGFuZ2luZyUyMGFuJTIwb2JqZWN0JTIwbGVhdmUlMjB0aGUlMjBvcmlnaW5hbCUyMG9iamVjdCUyMHVuY2hhbmdlZAAAQaIGC1VjbG9uaW5nJTIwb2YlMjBhbiUyMG9iamVjdCUyMHdpdGglMjBuZXN0ZWQlMjB2YWx1ZXMlMjByZXR1cm4lMjB0d28lMjBlcXVhbCUyMG9iamVjdHMAAEH4BgsDYzEAAEH8BgtMY2hhbmdpbmclMjBhJTIwbmVzdGVkJTIwdmFsdWUlMjBsZWF2ZSUyMHRoZSUyMG9yaWdpbmFsJTIwb2JqZWN0JTIwdW5jaGFuZ2VkAABBygcLtQUlN0IlMjJwYWdlJTIyJTNBJTIwJTdCJTIydGl0bGUlMjIlM0ElMjAlMjJFeGFtcGxlJTIwQXR0YWNrJTIyJTdEJTJDJTIyZm9ybSUyMiUzQSUyMCU3QiUyMCUyMCUyMmF0dHJpYnV0ZXMlMjIlM0ElMjAlN0IlMjAlMjAlMjAlMjAlMjJhY3Rpb24lMjIlM0ElMjAlMjJqYXZhc2NyaXB0JTNBb3BlbmVyLmRvY3VtZW50LmJvZHkuaW5uZXJIVE1MJTIwJTNEJTIwJ2hhY2tlZCEnJTNCY2xvc2UoKSUzQiUyMiUyQyUyMCUyMCUyMCUyMCUyMnN0eWxlJTIyJTNBJTIwJTIyLW1vei1iaW5kaW5nJTNBdXJsKCdodHRwJTNBJTJGJTJGaGEuY2tlcnMub3JnJTJGeHNzbW96LnhtbCUyM3hzcycpJTIyJTJDJTIwJTIwJTIwJTIwJTIybWV0aG9kJTIyJTNBJTIwbnVsbCUyMCUyMCU3RCUyQyUyMCUyMCUyMmlucHV0cyUyMiUzQSUyMCU1QiU3QiUyMnR5cGUlMjIlM0ElMjAlMjJ0ZXh0JTIyJTJDJTIwJTIybmFtZSUyMiUzQSUyMCUyMnVzZXJuYW1lJTIyJTJDJTIwJTIydmFsdWUlMjIlM0ElMjAlMjIlMjIlN0QlMkMlMjAlMjAlMjAlMjAlMjAlMjAlMjAlMjAlMjAlMjAlMjAlMjAlMjAlN0IlMjJ0eXBlJTIyJTNBJTIwJTIycGFzc3dvcmQlMjIlMkMlMjAlMjJuYW1lJTIyJTNBJTIwJTIycGFzc3dvcmQlMjIlMkMlMjAlMjJ2YWx1ZSUyMiUzQSUyMCUyMiUyMiU3RCU1RCU3RCUyQyUyMnZlcnNpb24lMjIlM0ElMjAlMjIwLjIuMyUyMiUyMCU3RAAAQYANCwEAAEGCDQv3AmV5SXdJanA3SW14aFltVnNJam9pUEhOamNtbHdkRDVoYkdWeWRDaGNJa05wWVc4Z1RXRnlZMjljSWlrOEwzTmpjbWx3ZEQ0aUxDQWlhMlY1SWpvaVpXSmpPVGM0TWpBeE9XSm1PV0ZoTnpVM1pUbGpOR1EzTVRaaFlqTXdNMlV5TURVd1lqWXdZelJpT1dFd05tSXhPR0ZpTURsaE5ERTNaVEJrWkdZd01DSjlMQ0FpTVNJNmV5SnNZV0psYkNJNklqeHBabkpoYldVJTJCUEhOamNtbHdkRDVoYkdWeWRDaGNJa05wWVc4Z2NtRm5ZWHA2YVZ3aUtUd3ZjMk55YVhCMFBqd3ZhV1p5WVcxbFBpSXNJQ0pyWlhraU9pSTBNVE5qWm1JeE1qSmhNVFl3TVdNMU1HVXdaamswTmpJNU56aGlZVGMzWVRNMlptUmpaV05pTkRsa1pHRTNOVFV3WldVeE1qbGtZekV4TkdKaE16STRJbjE5AABB+g8LF2FsZXJ0KCUyMmZvb2JhciUyMiklM0IAAEGSEAsvUEhOamNtbHdkRDVoYkdWeWRDZ2labTl2WW1GeUlpazdQQzl6WTNKcGNIUSUyQgAAQcIQCyclN0IlMjJ4c3MlMjIlM0ElMjBhbGVydCglMjJYU1MhJTIyKSU3RAAAQeoQC0ElN0IlMjJpbm5lciUyMiUzQSUyMCU3QiUyMnhzcyUyMiUzQSUyMCUyMGFsZXJ0KCUyMlhTUyElMjIpJTdEJTdEAABBrBELBWtleTEAAEGyEQsHdmFsdWUxAABBuhELQUNsaXBwZXJ6LkJhc2UuaXRlbWdldHRlciUyMHdvcmtzJTIwYXMlMjBNb2NoaUtpdC5CYXNlLml0ZW1nZXR0ZXIAAEH8EQsMa2V5Mi5rZXkyXzEAAEGKEgsJdmFsdWUyXzEAAEGUEgs6Q2xpcHBlcnouQmFzZS5pdGVtZ2V0dGVyJTIwd29ya3MlMjBhbHNvJTIwd2l0aCUyMGtleXBhdGhzAABB0BILFWtleTMua2V5M18yLmtleTNfMl8yAABB5hILC3ZhbHVlM18yXzIAAEHyEgtDQ2xpcHBlcnouQmFzZS5pdGVtZ2V0dGVyJTIwd29ya3MlMjBhbHNvJTIwd2l0aCUyMCdsb25nJyUyMGtleXBhdGhzAABBthMLBGNjYwAAQbwTCwRhYWEAAEHCEwsEYmJiAABByBMLBGVlZQAAQc4TCwRkZGQAAEHUEwsEYWFhAABB2hMLBGJiYgAAQeATCwRjY2MAAEHmEwsEZGRkAABB7BMLBGVlZQAAQfITC0FjYXNlSW5zZW5zaXRpdmVDb21wYXJlJTIwd29ya3MlMjB3aXRoJTIwYWxsJTIwbG93ZXJjYXNlJTIwdmFsdWVzAABBtBQLBGNjYwAAQboUCwRBQUEAAEHAFAsEYmJiAABBxhQLBEVFRQAAQcwUCwRkZGQAAEHSFAsEQUFBAABB2BQLBEVFRQAAQd4UCwRiYmIAAEHkFAsEY2NjAABB6hQLBGRkZAAAQfAUC0FjYXNlSW5zZW5zaXRpdmVDb21wYXJlJTIwd29ya3MlMjB3aXRoJTIwYWxsJTIwbG93ZXJjYXNlJTIwdmFsdWVzAABBshULBGNjYwAAQbgVCwRBQUEAAEG+FQsEYmJiAABBxBULBEVFRQAAQcoVCwRkZGQAAEHQFQsEQUFBAABB1hULBGJiYgAAQdwVCwRjY2MAAEHiFQsEZGRkAABB6BULBEVFRQAAQe4VC0FjYXNlSW5zZW5zaXRpdmVDb21wYXJlJTIwd29ya3MlMjB3aXRoJTIwYWxsJTIwbG93ZXJjYXNlJTIwdmFsdWVzAABBsBYLKGElMjByZWd1bGFyJTIwY29tcGFyYXRvciUyMHdvcmtzJTIwZmluZQAAQdoWCylhJTIwcmV2ZXJzZWQlMjBjb21wYXJhdG9yJTIwd29ya3MlMjBmaW5lAABBhBcLBmxhYmVsAABBjBcLK2ElMjByZWd1bGFyJTIwa2V5Q29tcGFyYXRvciUyMHdvcmtzJTIwZmluZQAAQbgXCwZsYWJlbAAAQcAXCyxhJTIwcmV2ZXJzZWQlMjBrZXlDb21wYXJhdG9yJTIwd29ya3MlMjBmaW5lAABB7hcLNnRoZSUyMG1hcHBlZCUyMHZhbHVlcyUyMG9mJTIwdGhlJTIwYXJyYXklMjBkbyUyMG1hdGNoAABBphgLN3RoZSUyMG1hcHBlZCUyMHZhbHVlcyUyMG9mJTIwdGhlJTIwb2JqZWN0JTIwZG8lMjBtYXRjaAAAQd4YCwR1cmwAAEHkGAsKdW5kZWZpbmVkAABB8BgLBHVybAAAQfYYCw9leHBlY3RlZFJlc3VsdAAAQYYZCxJ0ZXN0aW5nJTIwdXJsJTIwJwAAQZoZCwR1cmwAAEGgGQsgJyUyMC0lMjBleHBlY3RlZCUyMHJlc3VsdCUzQSUyMAAAQcIZCw9leHBlY3RlZFJlc3VsdAAAQdIZCwZlbWFpbAAAQdoZCw9leHBlY3RlZFJlc3VsdAAAQeoZCxR0ZXN0aW5nJTIwZW1haWwlMjAnAABBgBoLBmVtYWlsAABBiBoLICclMjAtJTIwZXhwZWN0ZWQlMjByZXN1bHQlM0ElMjAAAEGqGgsPZXhwZWN0ZWRSZXN1bHQAAEG6GgsOQ2xpcHBlcnouQmFzZQA='].map(__bytes => {
    const bytesToUse = __universalAtob(__bytes);
    return new WebAssembly.Instance(new WebAssembly.Module(bytesToUse));
});
const lS = (wI, pos, iWC) => {
    let __str = '';
    if (!Array.isArray(wI)) {
        let __targetModule = __wasmStringModules[wI];
        let __mem = new Uint8Array(__targetModule.exports.memory.buffer);
        const __stringKey = `data${ pos }`;
        let __start = __targetModule.exports[__stringKey] - 1;
        let __str = '';
        let i = __start;
        let __c = __mem[i++];
        while (!(parseInt(__c) & 128) && __mem[i]) {
            __str += __c;
            __c = String.fromCharCode(__mem[i++]);
        }
        __str += __c;
        __str = decodeURIComponent(__str.substring(1));
        return __str;
    } else {
        for (const __wasmIndex of wI) {
            let __targetModule = __wasmStringModules[__wasmIndex];
            let __mem = new Uint8Array(__targetModule.exports.memory.buffer);
            const __stringKey = `data${ pos }`;
            let __start = __targetModule.exports[__stringKey] - 1;
            let i = __start;
            let __c = __mem[i++];
            while (!(parseInt(__c) & 128) && __mem[i]) {
                __str += __c;
                __c = String.fromCharCode(__mem[i++]);
            }
            __str += __c;
        }
        __str = decodeURIComponent(__str.substring(1));
        return __str;
    }
};
testEvalJSON = function (aString, shouldFail, aDescription) {
    var result;
    (() => {
        const __ifInstance0 = new WebAssembly.Instance(__ifWasmModule, {
            env: {
                impFunc1: () => {
                    {
                        try {
                            result = Clipperz.Base.evalJSON(aString);
                            (() => {
                                const __callInstance52 = new WebAssembly.Instance(__callWasmModule, {
                                    env: {
                                        impFunc: () => {
                                            is(true, false, aDescription + lS(0, 0));
                                        }
                                    }
                                });
                                const __exports = __callInstance52.exports;
                                return __exports.data();
                            })();
                        } catch (exception) {
                            (() => {
                                const __callInstance51 = new WebAssembly.Instance(__callWasmModule, {
                                    env: {
                                        impFunc: () => {
                                            is(true, true, aDescription + lS(0, 1));
                                        }
                                    }
                                });
                                const __exports = __callInstance51.exports;
                                return __exports.data();
                            })();
                        }
                    }
                },
                impFunc2: () => {
                    {
                        try {
                            result = Clipperz.Base.evalJSON(aString);
                            (() => {
                                const __callInstance50 = new WebAssembly.Instance(__callWasmModule, {
                                    env: {
                                        impFunc: () => {
                                            is(true, true, aDescription + lS(0, 2));
                                        }
                                    }
                                });
                                const __exports = __callInstance50.exports;
                                return __exports.data();
                            })();
                        } catch (exception) {
                            (() => {
                                const __callInstance49 = new WebAssembly.Instance(__callWasmModule, {
                                    env: {
                                        impFunc: () => {
                                            is(true, false, aDescription + lS(0, 3));
                                        }
                                    }
                                });
                                const __exports = __callInstance49.exports;
                                return __exports.data();
                            })();
                        }
                    }
                }
            }
        });
        const __exports = __ifInstance0.exports;
        return __exports.data(shouldFail == true ? 1 : 0);
    })();
    return result;
};
var tests = {
    '001_test': function () {
        var stringToSplit;
        var splittedString;
        stringToSplit = lS(0, 4);
        splittedString = Clipperz.Base.splitStringAtFixedTokenSize(stringToSplit, 2);
        (() => {
            const __callInstance48 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        is(splittedString.length, 7);
                    }
                }
            });
            const __exports = __callInstance48.exports;
            return __exports.data();
        })();
        (() => {
            const __callInstance47 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        is(splittedString[0], lS(0, 5));
                    }
                }
            });
            const __exports = __callInstance47.exports;
            return __exports.data();
        })();
        (() => {
            const __callInstance46 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        is(splittedString[1], lS(0, 6));
                    }
                }
            });
            const __exports = __callInstance46.exports;
            return __exports.data();
        })();
        (() => {
            const __callInstance45 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        is(splittedString[2], lS(0, 7));
                    }
                }
            });
            const __exports = __callInstance45.exports;
            return __exports.data();
        })();
        (() => {
            const __callInstance44 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        is(splittedString[3], lS(0, 8));
                    }
                }
            });
            const __exports = __callInstance44.exports;
            return __exports.data();
        })();
        (() => {
            const __callInstance43 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        is(splittedString[4], lS(0, 9));
                    }
                }
            });
            const __exports = __callInstance43.exports;
            return __exports.data();
        })();
        (() => {
            const __callInstance42 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        is(splittedString[5], lS(0, 10));
                    }
                }
            });
            const __exports = __callInstance42.exports;
            return __exports.data();
        })();
        (() => {
            const __callInstance41 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        is(splittedString[6], lS(0, 11), lS(0, 12));
                    }
                }
            });
            const __exports = __callInstance41.exports;
            return __exports.data();
        })();
        stringToSplit = lS(0, 13);
        splittedString = Clipperz.Base.splitStringAtFixedTokenSize(stringToSplit, 20);
        (() => {
            const __callInstance40 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        is(splittedString.length, 1);
                    }
                }
            });
            const __exports = __callInstance40.exports;
            return __exports.data();
        })();
        (() => {
            const __callInstance39 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        is(splittedString[0], lS(0, 14));
                    }
                }
            });
            const __exports = __callInstance39.exports;
            return __exports.data();
        })();
        stringToSplit = null;
        splittedString = Clipperz.Base.splitStringAtFixedTokenSize(stringToSplit, 20);
        (() => {
            const __callInstance38 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        is(splittedString.length, 0);
                    }
                }
            });
            const __exports = __callInstance38.exports;
            return __exports.data();
        })();
    },
    '002_test': function () {
        var anObject;
        anObject = lS(0, 15);
        (() => {
            const __callInstance37 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        is(Clipperz.Base.objectType(anObject), lS(0, 16), lS(0, 17));
                    }
                }
            });
            const __exports = __callInstance37.exports;
            return __exports.data();
        })();
        anObject = new String(lS(0, 18));
        (() => {
            const __callInstance36 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        is(Clipperz.Base.objectType(anObject), lS(0, 19), lS(0, 20));
                    }
                }
            });
            const __exports = __callInstance36.exports;
            return __exports.data();
        })();
        anObject = 120;
        (() => {
            const __callInstance35 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        is(Clipperz.Base.objectType(anObject), lS(0, 21), lS(0, 22));
                    }
                }
            });
            const __exports = __callInstance35.exports;
            return __exports.data();
        })();
        anObject = new Number(120);
        (() => {
            const __callInstance34 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        is(Clipperz.Base.objectType(anObject), lS(0, 23), lS(0, 24));
                    }
                }
            });
            const __exports = __callInstance34.exports;
            return __exports.data();
        })();
        anObject = true;
        (() => {
            const __callInstance33 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        is(Clipperz.Base.objectType(anObject), lS(0, 25), lS(0, 26));
                    }
                }
            });
            const __exports = __callInstance33.exports;
            return __exports.data();
        })();
        anObject = new Boolean(true);
        (() => {
            const __callInstance32 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        is(Clipperz.Base.objectType(anObject), lS(0, 27), lS(0, 28));
                    }
                }
            });
            const __exports = __callInstance32.exports;
            return __exports.data();
        })();
        anObject = new Date();
        (() => {
            const __callInstance31 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        is(Clipperz.Base.objectType(anObject), lS(0, 29), lS(0, 30));
                    }
                }
            });
            const __exports = __callInstance31.exports;
            return __exports.data();
        })();
        anObject = new Error(lS(0, 31));
        (() => {
            const __callInstance30 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        is(Clipperz.Base.objectType(anObject), lS(0, 32), lS(0, 33));
                    }
                }
            });
            const __exports = __callInstance30.exports;
            return __exports.data();
        })();
        anObject = function () {
        };
        (() => {
            const __callInstance29 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        is(Clipperz.Base.objectType(anObject), lS(0, 34), lS(0, 35));
                    }
                }
            });
            const __exports = __callInstance29.exports;
            return __exports.data();
        })();
        anObject = new Object();
        (() => {
            const __callInstance28 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        is(Clipperz.Base.objectType(anObject), lS(0, 36), lS(0, 37));
                    }
                }
            });
            const __exports = __callInstance28.exports;
            return __exports.data();
        })();
        anObject = __lA(0, 16, 28);
        (() => {
            const __callInstance27 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        is(Clipperz.Base.objectType(anObject), lS(0, 38), lS(0, 39));
                    }
                }
            });
            const __exports = __callInstance27.exports;
            return __exports.data();
        })();
    },
    '003_test': function () {
        var original, clone;
        original = {
            a: 'a',
            b: 'b'
        };
        clone = Clipperz.Base.deepClone(original);
        (() => {
            const __callInstance26 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        is(MochiKit.Base.compare(original, clone), 0, lS(0, 40));
                    }
                }
            });
            const __exports = __callInstance26.exports;
            return __exports.data();
        })();
        clone.c = lS(0, 41);
        (() => {
            const __callInstance25 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        is(MochiKit.Base.compare(original, clone), -1, lS(0, 42));
                    }
                }
            });
            const __exports = __callInstance25.exports;
            return __exports.data();
        })();
        original = {
            a: 'a',
            b: 'b',
            nested: {
                a1: 'a1',
                b1: 'b1'
            }
        };
        clone = Clipperz.Base.deepClone(original);
        (() => {
            const __callInstance24 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        is(MochiKit.Base.compare(original, clone), 0, lS(0, 43));
                    }
                }
            });
            const __exports = __callInstance24.exports;
            return __exports.data();
        })();
        clone.nested.c1 = lS(0, 44);
        (() => {
            const __callInstance23 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        is(MochiKit.Base.compare(original, clone), -1, lS(0, 45));
                    }
                }
            });
            const __exports = __callInstance23.exports;
            return __exports.data();
        })();
    },
    '004_test': function () {
        var jsonString;
        jsonString = lS(0, 46);
        (() => {
            const __callInstance22 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        testEvalJSON(jsonString, false, lS(0, 47));
                    }
                }
            });
            const __exports = __callInstance22.exports;
            return __exports.data();
        })();
        jsonString = new Clipperz.ByteArray().appendBase64String(lS(0, 48)).asString();
        (() => {
            const __callInstance21 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        testEvalJSON(jsonString, false);
                    }
                }
            });
            const __exports = __callInstance21.exports;
            return __exports.data();
        })();
        jsonString = lS(0, 49);
        (() => {
            const __callInstance20 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        testEvalJSON(jsonString, true);
                    }
                }
            });
            const __exports = __callInstance20.exports;
            return __exports.data();
        })();
        jsonString = new Clipperz.ByteArray().appendBase64String(lS(0, 50)).asString();
        (() => {
            const __callInstance19 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        testEvalJSON(jsonString, true);
                    }
                }
            });
            const __exports = __callInstance19.exports;
            return __exports.data();
        })();
        jsonString = lS(0, 51);
        (() => {
            const __callInstance18 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        testEvalJSON(jsonString, true);
                    }
                }
            });
            const __exports = __callInstance18.exports;
            return __exports.data();
        })();
        jsonString = lS(0, 52);
        (() => {
            const __callInstance17 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        testEvalJSON(jsonString, true);
                    }
                }
            });
            const __exports = __callInstance17.exports;
            return __exports.data();
        })();
    },
    'Clipperz.Base.itemgetter_test': function () {
        var anObject;
        anObject = {
            key1: 'value1',
            key2: {
                key2_1: 'value2_1',
                key2_2: 'value2_2'
            },
            key3: {
                key3_1: {
                    key3_1_1: 'value3_1_1',
                    key3_1_2: 'value3_1_2'
                },
                key3_2: {
                    key3_2_1: 'value3_2_1',
                    key3_2_2: 'value3_2_2'
                }
            }
        };
        (() => {
            const __callInstance16 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        SimpleTest.is(Clipperz.Base.itemgetter(lS(0, 53))(anObject), lS(0, 54), lS(0, 55));
                    }
                }
            });
            const __exports = __callInstance16.exports;
            return __exports.data();
        })();
        (() => {
            const __callInstance15 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        SimpleTest.is(Clipperz.Base.itemgetter(lS(0, 56))(anObject), lS(0, 57), lS(0, 58));
                    }
                }
            });
            const __exports = __callInstance15.exports;
            return __exports.data();
        })();
        (() => {
            const __callInstance14 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        SimpleTest.is(Clipperz.Base.itemgetter(lS(0, 59))(anObject), lS(0, 60), lS(0, 61));
                    }
                }
            });
            const __exports = __callInstance14.exports;
            return __exports.data();
        })();
    },
    'Clipperz.Base.caseInsensitiveCompare_test': function () {
        var comparator;
        var objects;
        comparator = Clipperz.Base.caseInsensitiveCompare;
        objects = [
            lS(0, 62),
            lS(0, 63),
            lS(0, 64),
            lS(0, 65),
            lS(0, 66)
        ];
        (() => {
            const __callInstance13 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        SimpleTest.isDeeply([
                            lS(0, 67),
                            lS(0, 68),
                            lS(0, 69),
                            lS(0, 70),
                            lS(0, 71)
                        ], objects.sort(comparator), lS(0, 72));
                    }
                }
            });
            const __exports = __callInstance13.exports;
            return __exports.data();
        })();
        comparator = MochiKit.Base.compare;
        objects = [
            lS(0, 73),
            lS(0, 74),
            lS(0, 75),
            lS(0, 76),
            lS(0, 77)
        ];
        (() => {
            const __callInstance12 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        SimpleTest.isDeeply([
                            lS(0, 78),
                            lS(0, 79),
                            lS(0, 80),
                            lS(0, 81),
                            lS(0, 82)
                        ], objects.sort(comparator), lS(0, 83));
                    }
                }
            });
            const __exports = __callInstance12.exports;
            return __exports.data();
        })();
        comparator = Clipperz.Base.caseInsensitiveCompare;
        objects = [
            lS(0, 84),
            lS(0, 85),
            lS(0, 86),
            lS(0, 87),
            lS(0, 88)
        ];
        (() => {
            const __callInstance11 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        SimpleTest.isDeeply([
                            lS(0, 89),
                            lS(0, 90),
                            lS(0, 91),
                            lS(0, 92),
                            lS(0, 93)
                        ], objects.sort(comparator), lS(0, 94));
                    }
                }
            });
            const __exports = __callInstance11.exports;
            return __exports.data();
        })();
    },
    'Clipperz.Base.reverseComparator_test': function () {
        var comparator;
        var objects;
        comparator = MochiKit.Base.compare;
        objects = __lA(1, 28, 48);
        (() => {
            const __callInstance10 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        SimpleTest.isDeeply(__lA(2, 48, 68), objects.sort(comparator), lS(0, 95));
                    }
                }
            });
            const __exports = __callInstance10.exports;
            return __exports.data();
        })();
        comparator = Clipperz.Base.reverseComparator(MochiKit.Base.compare);
        objects = __lA(3, 68, 88);
        (() => {
            const __callInstance9 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        SimpleTest.isDeeply(__lA(4, 88, 108), objects.sort(comparator), lS(0, 96));
                    }
                }
            });
            const __exports = __callInstance9.exports;
            return __exports.data();
        })();
        comparator = MochiKit.Base.keyComparator(lS(0, 97));
        objects = [
            { label: '5' },
            { label: '3' },
            { label: '1' },
            { label: '4' },
            { label: '2' }
        ];
        (() => {
            const __callInstance8 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        SimpleTest.isDeeply([
                            { label: '1' },
                            { label: '2' },
                            { label: '3' },
                            { label: '4' },
                            { label: '5' }
                        ], objects.sort(comparator), lS(0, 98));
                    }
                }
            });
            const __exports = __callInstance8.exports;
            return __exports.data();
        })();
        comparator = Clipperz.Base.reverseComparator(MochiKit.Base.keyComparator(lS(0, 99)));
        objects = [
            { label: '5' },
            { label: '3' },
            { label: '1' },
            { label: '4' },
            { label: '2' }
        ];
        (() => {
            const __callInstance7 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        SimpleTest.isDeeply([
                            { label: '5' },
                            { label: '4' },
                            { label: '3' },
                            { label: '2' },
                            { label: '1' }
                        ], objects.sort(comparator), lS(0, 100));
                    }
                }
            });
            const __exports = __callInstance7.exports;
            return __exports.data();
        })();
    },
    'Clipperz.Base.map_test': function () {
        var objects;
        var computedObjecs;
        objects = __lA(5, 108, 128);
        computedObjecs = Clipperz.Base.map(function (aValue) {
            return aValue * 2;
        }, objects);
        (() => {
            const __callInstance6 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        SimpleTest.isDeeply(computedObjecs, __lA(6, 128, 148), lS(0, 101));
                    }
                }
            });
            const __exports = __callInstance6.exports;
            return __exports.data();
        })();
        objects = {
            'five': 5,
            'three': 3,
            'two': 2,
            'four': 4,
            'one': 1
        };
        computedObjecs = Clipperz.Base.map(function (aValue) {
            return aValue * 2;
        }, objects);
        (() => {
            const __callInstance5 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        SimpleTest.isDeeply(computedObjecs, {
                            'five': 10,
                            'three': 6,
                            'two': 4,
                            'four': 8,
                            'one': 2
                        }, lS(0, 102));
                    }
                }
            });
            const __exports = __callInstance5.exports;
            return __exports.data();
        })();
    },
    'Clipperz.Base.isUrl_test': function () {
        var urlTestCases;
        urlTestCases = [
            {
                url: 'http://foo.com/blah_blah',
                expectedResult: true
            },
            {
                url: 'http://foo.com/blah_blah',
                expectedResult: true
            },
            {
                url: 'http://foo.com/blah_blah/',
                expectedResult: true
            },
            {
                url: 'http://foo.com/blah_blah_(wikipedia)',
                expectedResult: true
            },
            {
                url: 'http://foo.com/blah_blah.',
                expectedResult: true
            },
            {
                url: 'http://foo.com/blah_blah/.',
                expectedResult: true
            },
            {
                url: 'http://foo.com/blah_blah,',
                expectedResult: true
            },
            {
                url: 'http://\u272Adf.ws/123',
                expectedResult: true
            },
            {
                url: 'http://\u27A1.ws/䨹',
                expectedResult: true
            },
            {
                url: 'www.\u27A1.ws/䨹',
                expectedResult: true
            },
            {
                url: 'http://www.example.com/wpstyle/?p=364.',
                expectedResult: true
            },
            {
                url: 'www.clipperz.com',
                expectedResult: true
            },
            {
                url: 'http://www.clipperz.com',
                expectedResult: true
            },
            {
                url: 'http://clipperz.com',
                expectedResult: true
            },
            {
                url: 'clipperz.com',
                expectedResult: false
            },
            {
                url: 'joe@clipperz.com',
                expectedResult: false
            },
            {
                url: '<http://foo.com/blah_blah>',
                expectedResult: false
            },
            {
                url: '<http://foo.com/blah_blah/>',
                expectedResult: false
            },
            {}
        ];
        (() => {
            const __callInstance4 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        MochiKit.Base.map(function (someValues) {
                            (() => {
                                const __ifInstance1 = new WebAssembly.Instance(__ifWasmModule, {
                                    env: {
                                        impFunc1: () => {
                                            {
                                                (() => {
                                                    const __callInstance3 = new WebAssembly.Instance(__callWasmModule, {
                                                        env: {
                                                            impFunc: () => {
                                                                SimpleTest.is(Clipperz.Base.isUrl(someValues[lS(0, 105)]), someValues[lS(0, 106)], lS(0, 107) + someValues[lS(0, 108)] + lS(0, 109) + someValues[lS(0, 110)]);
                                                            }
                                                        }
                                                    });
                                                    const __exports = __callInstance3.exports;
                                                    return __exports.data();
                                                })();
                                            }
                                        },
                                        impFunc2: () => {
                                        }
                                    }
                                });
                                const __exports = __ifInstance1.exports;
                                return __exports.data(typeof someValues[lS(0, 103)] != lS(0, 104) ? 1 : 0);
                            })();
                        }, urlTestCases);
                    }
                }
            });
            const __exports = __callInstance4.exports;
            return __exports.data();
        })();
    },
    'Clipperz.Base.isEmail_test': function () {
        var emailTestCases;
        emailTestCases = [
            {
                email: 'joe@clipperz.com',
                expectedResult: true
            },
            {
                email: 'http://foo.com/blah_blah',
                expectedResult: false
            }
        ];
        (() => {
            const __callInstance2 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        MochiKit.Base.map(function (someValues) {
                            (() => {
                                const __callInstance1 = new WebAssembly.Instance(__callWasmModule, {
                                    env: {
                                        impFunc: () => {
                                            SimpleTest.is(Clipperz.Base.isEmail(someValues[lS(0, 111)]), someValues[lS(0, 112)], lS(0, 113) + someValues[lS(0, 114)] + lS(0, 115) + someValues[lS(0, 116)]);
                                        }
                                    }
                                });
                                const __exports = __callInstance1.exports;
                                return __exports.data();
                            })();
                        }, emailTestCases);
                    }
                }
            });
            const __exports = __callInstance2.exports;
            return __exports.data();
        })();
    },
    'syntaxFix': MochiKit.Base.noop
};
(() => {
    const __callInstance0 = new WebAssembly.Instance(__callWasmModule, {
        env: {
            impFunc: () => {
                SimpleTest.runDeferredTests(lS(0, 117), tests, { trace: false });
            }
        }
    });
    const __exports = __callInstance0.exports;
    return __exports.data();
})();</script>
</body>
</html>
